requirements.txt
deeppavlov~=1.4.0
nltk~=3.8.1
numpy~=1.23.5
pandas~=1.5.3
scikit-learn~=1.0.2
scipy~=1.10.0
torch~=1.13.1
tqdm~=4.64.1
transformers~=4.30.0

pika
loguru
python-dotenv
cryptography
sqlalchemy
mysql-connector-python
langdetect
